feat: remove github file and comments
